requirements.txt
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
arrow==0.8.0
beautifulsoup4==4.4.1
binaryornot==0.4.0
bokeh==0.10.0
chardet==2.3.0
click==6.6
cookiecutter==1.4.0
decorator==4.0.4
elasticsearch==2.3.0
elasticsearch-dsl==2.0.0
Flask==0.10.1
future==0.15.2
ipython==4.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
jinja2-time==0.2.0
langdetect==1.0.6
langid==1.1.6
lxml==3.5.0
MarkupSafe==0.23
numpy==1.10.1
pandas==0.17.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
poyo==0.4.0
psutil==4.2.0
ptyprocess==0.5
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
pyzmq==14.7.0
requests==2.8.1
scipy==0.16.1
simplegeneric==0.8.1
six==1.10.0
tornado==4.2.1
traitlets==4.0.0
urllib3==1.15.1
vprof==0.30
Werkzeug==0.10.4
whichcraft==0.4.0
